+++ client/tictactoe/.prettierrc.yaml
