+++ client/tictactoe/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tictactoe-client",
+  "name": "tictactoe",
   "version": "0.1.0",
   "private": true,
   "scripts": {
