+++ client/tictactoe/src/assets/logo.png
