+++ server/tictactoe.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.452
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "tictactoe-server", "tictactoe-server\tictactoe-server.csproj", "{FB01B454-CCB0-4969-8231-635CEBB01393}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TicTacToe", "TicTacToe\TicTacToe.csproj", "{FB01B454-CCB0-4969-8231-635CEBB01393}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
