+++ server/tictactoe/appsettings.json
@@ -1,6 +1,6 @@
 {
 	"ConnectionStrings": {
-		"DefaultConnection": "Data Source=Tictactoe.db"
+		"DefaultConnection": "Data Source=tictactoe.db"
 	},
 	"Logging": {
 		"LogLevel": {
